[AAP-9881] Update project import logic
Before the project was created as a result of the import task
completion.

This PR changes the import workflow. Project create
request handler creates a project record in the database
in `pending` state and spawns a background job.

The background job will update project data and set the `import_state`
to `completed` if the import was successfull or `failed` otherwise.

Implements: https://issues.redhat.com/browse/AAP-9881