Fix broken unit tests (#161)
The following PR was merged to main in order to initialize permissions
for RBAC data. Unit tests for roles endpoints that merged earlier in the
day was initializing permission data in order to test. This cause some
duplicate values in the DB and due to unique constraints resulted in
failures by error. This PR takes the initialization into account and
fixes the tests.

PR for initializing permissions data:
[https://github.com/ansible/aap-eda/pull/148](https://github.com/ansible/aap-eda/pull/148)