Update docker-compose-stage (#214)
That would allow us to run a deployment able to integrate with AWX and
rootless podman in linux hosts.
Includes:

- Removes the "url" from the name of the image_url envvar. 
- Adds configuration to mount the podman socket. (socket path variables
+ run as root for rootless podman by default)
- Adds configuration for AWX integration
- Updates docs accordingly

docker-compose-dev is not covered as we need to investigate another
approach because changing the user_id would break the bind mounts. Also
we have issues with mac as the support to mount the sockets there is
problematic. We have the plan to add a new composer file for mac that
overrides the worker conf with proper instructions. Discussed with
@mkanoor

---------

Co-authored-by: Tom Tuffin <71447672+ttuffin@users.noreply.github.com>