Support activation restart policy
Add delayed queue task to monitor the status of activation
according to its restart policy. Restart a activation if the status
is failed and the policy is on-failure, or the status is completed
(meaning the activation succeeded, rulesets should running ok) but
last received heartbeat was too long ago and the restart policy is
always.

This implementation depends on the rulebook engine to send a heart-
beat event at a fixed interval.

Implements AAP-11337: Implement Rulebook activation restart policy