[AAP-11784] Fixed podman credential issues on linux
On Linux it seems like podman needs a credential file and doesn't
accept credentials on the api call like it does on Mac.
The API version is also lower.
I have made changes in this PR to create the auth.json file with
credentials which match the docker auth config file format

https://issues.redhat.com/browse/AAP-11784