[AAP-11920] Implement custom API exception handler
Include error code in API response.

The API returns 403 status code in case when user is not
authenticated and when user has no permission to perform an
action. In order to disambiguate response from API server,
this pull request adds a new field `code` in error response body.

For stability the code is only filled for limited number exceptions classes
related to the user authentication:

  - `AuthenticationFailed`
  - `NotAuthenticated`
  - `PermissionDenied`