fix: send the Hello message after the DB is updated
When the server side receives a websocket messages each message
is handled by a separate asyncio task. In the task the first thing
we do is send a "Hello" message to the ansible-rulebook. If ansible
rulebook terminates before this Hello message gets to its the asyncio
task aborts and doesn't write any messages into the DB. This leads
to messages getting lost.
With this fix now we send the "Hello" message only after we have
written the data to the DB.

https://issues.redhat.com/browse/AAP-16264