fix(ci): Task queued execution in testing
In an environment with a default worker running, tests that enqueue tasks on the default queue  result in the task being executed by the default worker within the default worker's environment.  This causes a cross-environment issue for tests (e.g., test_projects) which depend on the executing task accessing the test environment's DB which results in test failure.  This change adds a test environment-specific queue (still referenced as 'default_queue' thereby providing a logically consistent facade) on which these tasks are run within the same environment as the test's DB.

Signed-off-by: Joe Shimkus <jshimkus@redhat.com>