chore: allow for minikube deployments to use quay images
Previously the minikube workflow required the UI and eda server
images to be built locally before they can be used.
Which is cumbersome, this PR adds a new task called
 * quay-deploy

which allows us to run directly from quay images and it also allows
for the images to be overridden for UI and eda server local images
using the following env vars

* EDA_UI_IMAGE=localhost/eda-ui:latest
* EDA_UI_IMAGE_PULL_POLICY=Never
* EDA_IMAGE=localhost/eda-server:latest
* EDA_IMAGE_PULL_POLICY=Never

To run the minikube deployment run the following command
* task minikube:quay-deploy
Once all the pods are up and running
* task minikube:fp:ui
To cleanup
* task minikube:clean

Given that we might be switching to use the ansible-playbook
deployment this is a short term solution.