fix: check the duplicates in extra var only during create
When we create an activation for the first time, we update the
extra vars with keys from the credentials. This check should
happen only when the Activation is being created. We should not
do the extra var key check during post activation otherwise
we will always see the duplicates which are the keys we added
during create.