fix: [AAP-23378] add PENDING_WORKERS_OFFLINE activation status
This provides the agreed upon distinction between
- PENDING
    a normal occurrence of a request not yet running
- PENDING_WORKERS_OFFLINE
    an abnormal occurrence where the request has not yet run and cannot run due
    to lack of workers; note that this includes auto-start events where we
    definitively know the request's state
- WORKERS_OFFLINE
    an abnormal occurrence where we do not know the request's state

PENDING_WORKERS_OFFLINE maintains the fail-over functionality previously dependent
solely on WORKERS_OFFLINE.

For consistency with other log messages from the processing code the content of the message indicating issues with the workers has been modified from that documented in AAP-23378.  It is worded as follows:

    The workers available to process request type {request_type} for
    {process_parent_type} {process_parent_id} are failing liveness checks.
    There may be an issue with the node; please contact the administrator.

Orchestrator log messages were updated due to previous change that specifically set
the request type if none was specified.