chore: avoid pytest conflicting with deployed eda-server
This change uses a separate redis DB from the default when running
the eda-server pytests in an effort to avoid stepping on the DB used
in an eda-server deployment.  It's not a guarantee in that a deployment
may be differently configured, but is sufficient in a general development
scenario.

Also, removed a redundant copy of the pytest fixture redis_paramters.