chore: undo the migration check
The migration check was added in PR #957 to make podman-compose
handle dependent containers coming up before the DB was ready.
This check causes podman-compose/podman to terminate containers
if they dont respond within 10 seconds. So reverting for now
till we can come up with other solution to solve this for
podman-compose