chore: fix dependency for podman
All of our workers have a dependency on the api-worker which
has dependency to postgres and redis.
We dont need to have default worker and activation workers have
a direct dependency to postgres and redis. They can have a direct
dependency to api-worker.

In podman with the --requires if we have a direct dependency to
redis and postgres from default and activation workers they might
come up before even redis is queued up and break the dependency chain.
Causing them to not start.

```
[eda-scheduler]         | Error: unable to start container c979979377dd402dacee7856328c205810362bf3abdb5c85d0e291ecb2ca6e45: generating dependency graph for container c979979377dd402dacee7856328c205810362bf3abdb5c85d0e291ecb2ca6e45: container 4f8bfe666edea37dd185d51066e09d5fcb512db3dcd2608f3acc2b5f106f9ad9 depends on container 92cc8f498d3dba3a99ca1b649fa62ba03fd5663891dc32217cd79ad2573e2315 not found in input list: no such container
[eda-activation-worker] | Error: unable to start container 2edd9cab6bcf841db30689f6379a992b0e2ecb6b7befb3a5eca9d77f945f89f8: generating dependency graph for container 2edd9cab6bcf841db30689f6379a992b0e2ecb6b7befb3a5eca9d77f945f89f8: container 4f8bfe666edea37dd185d51066e09d5fcb512db3dcd2608f3acc2b5f106f9ad9 depends on container 92cc8f498d3dba3a99ca1b649fa62ba03fd5663891dc32217cd79ad2573e2315 not found in input list: no such container
```