fix: remove organization default from BaseOrgModel
The organization is a required parameter everywhere it appears in the API.
Having a default for it runs counter to this and was only of value internally
to EDA for eliminating the need to specify an organization when creating test
instances of various models.  This introduced a dependency on the default
organization that was not applicable to EDA's normal operation and resulted in
the generation of faults if the name of the organization was changed.

This change modifies EDA's unit testing to use the pre-existing pytest fixture
'default_organization' in testing rather than a functionally unnecessary and
fragile internal convenience.

This change does not completely eliminate the presumption of the existence of a
default organization accessible by a specific name. This presumption, however,
is restricted to the initial database population run once when an EDA
deployment is started for the first time.