fix: update activation instance's started_at
The started_at used to be the instance created time, which is even
before the DE is downloaded. The download time plus the container
init time can exceed the default 30 seconds LIVENESS timeout window.
This fix resets the started_at time to be at the moment after the
container is initiated.