feat: handle binary files
For Kerberos we need to support a keytab file which is a binary
file and its used for service accounts during authentication
with Kerberos.

This uses the format: binary_base64 in the field definition to
let the UI know that the expected file is supposed to be binary
and needs to be sent up as base64
You can create a credential with the following schema

Inputs configuration

```
fields:
  - id: keytab
    type: string
    label: Kerberos Keytab file
    format: binary_base64
    secret: true
    help_text: Please select a Kerberos Keytab file
    multiline: true
```

Injector Configuration

```
file:
  template.keytab_file: '{{ keytab }}'
```

No changes are needed to ansible-rulebook it already handles
binary files.

https://issues.redhat.com/browse/AAP-25518