fix: injected env referencing to injected file
This fix solves the exception when creating a credential_type
with injected envs referencing to injected files, and the
exception when creating a credential based on such types.

The root cause is that {{ eda.filename.abc }} is not substituable.
We need to make this allowable during validation and skip substitution
for non extra_vars at credential creation time.

Related to AAP-35595