[AAP-39182] - Ensure DE validation errors display under text box (#1237)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
Decision environment image validation errors previously did not show up
under the fields text box in the UI as expected. Instead they displayed
in a banner at the top of the page.

This PR ensures that we return the field name along with the error where
necessary, so that the error displays under the text box as expected.

[AAP-39182](https://issues.redhat.com/browse/AAP-39182)
<!-- If applicable, provide a link to the issue that is being addressed
-->

<!-- What is being changed? -->
<!-- Why is this change needed? -->
<!-- How does this change address the issue? -->
<!-- Does this change introduce any new dependencies, blockers or
breaking changes? -->
<!-- How it can be tested? -->