fix: discard null secret values when encrypting AAP-42205 (#1239)
In a deployment configured with mTLS postgres the activation can not be
created returning an Unhandled 500 error. The error happens when calling
ansible-vault for encryption of values to be sent to ansible-rulebook.
If a value is None is sent to Pexpect which expects a string. In this
case the db_password is None since it is configured with mTLS.
Jira: https://issues.redhat.com/browse/AAP-42205

Also includes:
- Encryption under a try/except block since it performs multiple calls
that might raise a variety of exceptions without context.
- Update the docker-compose-mtls and add the missing certs for OoO local
deployment