fix: process k8s_service_name during activation update (#1243)
Fix an internal 500 error when calling api to update an activation on
k8s

[AAP-42437](https://issues.redhat.com/browse/AAP-42437)

If no new name or k8s_service_name is provided when request updating the
activation, the values from the existing activation are pulled. They are
needed for validation.
In addition, when the k8s_service_name is cleared through the update
request, it will be regenerated with the default value.

To test the fix
1. Update an activation without changing name on k8s, it no longer
raises a 500 internal error.
2. Create an activation with a customized k8s_service_name and clear it
through an update, verify the activation's k8s_service_name is assigned
to a default value generated based on the activation name.