chore: update docker-compose-dev to set multiple workers AAP-42483 (#1245)
The purpose of this PR is to set multiple default workers and default
activation workers per node to simulate the scenario that we will have
in OCP with multiple main processes of dispatcher listening in the same
channel.
I don't use `replicas` of the compose spec because it has some
limitations, like for example the capability of opening a port, so it is
better a little copy-paste of the number of desired services.
Jira: https://issues.redhat.com/browse/AAP-42483

Signed-off-by: Alex <aizquier@redhat.com>