[AAP-43604] fix: properly detect whether source_mappings are empty (#1263)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
When the even streams are removed from an activation, the UI sends
"[]\n" to the backend which were interpreted non-empty thus the event
stream credential was mistakenly added.
<!-- If applicable, provide a link to the issue that is being addressed
-->
[AAP-43604](https://issues.redhat.com/browse/AAP-43604)
<!-- What is being changed? -->
Parse the source_mapping string and accurately determine whether there
are source mapping or empty. Only inject the extra_vars when there are
true source mappings.