fix: ensure database settings printed in startup logging (#1316)
We have detected issues with the current code in startup logging where
the database settings are not printed. The current code has a bug where
we assumed the databases object would be a `box` object provided by
dynaconf, but it is actually a dict, so it never will have attributes.
This code fixes the issue and ensures the database settings are printed.

Signed-off-by: Alex <aizquier@redhat.com>