fix: Run noop loop in scheduler under dispatcherd flag (#1332)
This PR changes the approach to handle the dispatcherd flag in the
scheduler. Instead or returning an error and creating noise for users,
that also need to modify their deployments to not deploy the scheduler
which would be restarting all the time in default deployments, it
applies a less disruptive approach by just warning about it and running
a noop loop.

Signed-off-by: Alex <aizquier@redhat.com>