fix: run advisory_lock import/sync project in the right place AAP-47111 (#1333)
Fix a bug introduced in https://github.com/ansible/eda-server/pull/1296
where we are running under the advisory lock not the real import/sync
task but the proxy that schedules the job for rq and dispatcherd.

This fix run the task under the lock in the right place. Also allows to
merge both test_job_uniqueness and test_project_job_uniqueness tests

Jira: AAP-47111

Signed-off-by: Alex <aizquier@redhat.com>