ci: make CI github workflow reusable (#1339)
Convert the CI workflow in reusable workflow, called by run CI workflow
with the existing logic. This allows to other forks and mirrors to reuse
the CI pipeline with a different trigger configuration.

This change needs too to modify the status checks required in the repo
configuration after merge, It can not be done before because they don't
exist yet. In this PR you can see the new ones executed.

Signed-off-by: Alex <aizquier@redhat.com>