test: fix flaky tests caused by multithreaded tests (#1334)
We have sporadic unexpected failures in tests. I think this is caused by
`test_job_uniqueness` which uses multithreading and seems to interfere
intermittently with other mocks. Since I could not found a way to avoid
it or a better way to do this test (which I think right now is clean and
reliable except for the mock interference) I think the best way is to
just keep them separately, which is a good and common pattern in
general.

It marks the test with a new marker "multithreaded" which is never
executed by default. Then CI stuff is updated to execute both, default
and multithreaded, separately.

Depends on https://github.com/ansible/eda-server/pull/1333

Signed-off-by: Alex <aizquier@redhat.com>