Revert "ci: make CI github workflow reusable (#1339)" (#1341)
The approach of #1339 is not enough for what we need to do. We need to
support auth and private sources in forks and mirrors, so it means, we
need to alter the workflow itself, not only the call to the workflow.
Also it has a problem, because now the workflow is called, there is a
task that is failing in main because it has no acccess to some org
variables
([example](https://github.com/ansible/eda-server/actions/runs/15472565707/job/43560431636#step:12:24))
I did not noticed this problem because it is only executed on main.

This PR reverts the change and restore the previous configuration. It
will need too to revert the status checks required in the repo
configuration after merge.

After this rollback I will create a new PR with a different approach:
Instead of making the whole workflow reusable I will extract some jobs
into reusable actions, this allow us to compose different workflows by
reusable pieces and customize steps, like configuring ssh/tokens for
private sources and so on. This PR will have also in consideration the
variables that are required for some steps.