feat: Not allow local resource management by default (#1359)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
AAP-48659 asks to remove ALLOW_LOCAL_RESOURCE_MANAGEMENT in EDA. This
flag is no longer used across AAP.
<!-- If applicable, provide a link to the issue that is being addressed
-->
https://issues.redhat.com/browse/AAP-48659
<!-- What is being changed? -->
This PR converts ALLOW_LOCAL_RESOURCE_MANAGEMENT to an internal flag.
The default value is False meaning EDA is not allowed to create/modify
shared resources. PDE does not need to manipulate the flag. The flag is
not completely removed for development purpose. It can be turn on to run
EDA standalone.
<!-- Why is this change needed? -->
<!-- How does this change address the issue? -->
<!-- Does this change introduce any new dependencies, blockers or
breaking changes? -->
<!-- How it can be tested? -->