chore(dispatcherd): parameterize and increase queue healthcheck timeout (#1360)
In some scenarios with high concurrent workload I have observed
intermittent healthcheck failures. This PR's increases the default
timeout with a more safe value (default is 1) and exposes it as setting.
Depending on the setup this value might need to be adjusted but scaling
up workers or adjusting subprocesses per worker should be the right
approach in the most of the cases.

Signed-off-by: Alex <aizquier@redhat.com>