fix: moving project out of stuck pending or running state (#1381)
<!-- Mandatory: Provide a clear, concise description of the changes and
their purpose -->
Fix bug in existing monitor_project_tasks for pending state. 
Add new logic to handle running state. 
Add new django command to manually change the import state of a project
<!-- If applicable, provide a link to the issue that is being addressed
-->
https://issues.redhat.com/browse/AAP-51139
<!-- What is being changed? -->
Fix the bug in requeueing pending projects that are no longer in the
queue.
Detect running projects that are no longer in the queue or already
failed but not updated in database.
The django command can be used as a last resort helping the project to
move out of a stuck state.
<!-- Why is this change needed? -->
Once a project is stuck in pending or running, it cannot be restarted.
This work helps to move such projects out of the situation.
<!-- How does this change address the issue? -->
<!-- Does this change introduce any new dependencies, blockers or
breaking changes? -->
<!-- How it can be tested? -->