fix: when doing credential test inputs and metadata are mandatory (#1384)
When doing external sms credential test we need both inputs and
metadata.

There is no UI field related to this unlike other ValidationErrors We
return a 400 when either of them is missing

https://issues.redhat.com/browse/AAP-51130

When the test is run through the UI it adheres by the schema and ensures
that all required fields are identified in red and only once the user
has provided all the parameters is the test run. When the same test is
run via scripts or collection these checks are not performed so there is
a chance they might pass in an empty dictionary this PR basically
proectes the backend from these cases.

<img width="1089" height="490" alt="Screenshot 2025-08-20 at 9 49 19 PM"
src="https://github.com/user-attachments/assets/f7696ee1-b3a4-427c-8e32-27037a5e9f04"
/>