fix: the test function was broken for eda_credential
The test function for eda_credential was broken, it needs
to support some of the Credential Types that have optional
metadata fields like GitHub App Installation Access Token Lookup
We need to support empty metadata during the test.
If the metadata is empty we have to look at the credentials schema
and see if it has mandatory fields and then raise an error.
If the fields are optional then we should allow empty metadata

https://issues.redhat.com/browse/AAP-52232