fix: parallelize queue_dispatching in monitor loop - AAP-37345 (#1357)
Try to solve the issue of the bottleneck that supposes the superlinear
complexity of the current sequential and no scalable monitor_rulebook
loop.
Each user request and monitor needs to be queue-dispatched before being
processed by the manager. I follow the same approach as for the rest of
tasks by implementing a "lock" and "no_lock" versions of
`queue_dispatch` and scheduling it per every user and monitor request in
the monitor for the default worker. In this way we can parallelize the
queue_dispatcher while preserving the consistency, leaving the current
monitor loop as a lightweight iterator that schedules all the subsequent
tasks. The monitor can scale better in big clusters by just increasing
the number of workers.

Jira: https://issues.redhat.com/browse/AAP-37345

---------

Signed-off-by: Alex <aizquier@redhat.com>