Fix downstream failure from DAB (#1398)
Failure documented in:

https://github.com/ansible/django-ansible-base/pull/852

I expect that DAB is seeing `test_failed_decrypt` fail because of minor
differences in behavior due to versions of dependencies. The failure
seems to occur due to minor and inconsequential differences in the exact
syntax of the error message observed. This change generalizes the
expected regex for the exception's text.