Move to Runtime Platform Flags (#1407)
Jira: https://issues.redhat.com/browse/AAP-45879

This PR moves the EDA feature flag source from Settings based feature
flags, to platform database level feature flags.
This also makes gateway the provider of this data for the platform.

This is required to enable platform level feature flags.

Gateway is the provider of the feature flag value updates, and is
resource synced to EDA/other components.

It can be validated and tested simply with AAP Dev, and requires the
following DAB PR to be merged in beforehand -
https://github.com/ansible/django-ansible-base/pull/875
Once the DAB PR is merged, I will update this PR to point back to the
upstream devel branch of DAB

To test -

1. Clone AAP-Dev
2. make configure-sources
3. Select DAB, GW, and EDA for sources
4. make aap
5. Attempt to hit <GATEWAY_API>/feature_flags/X api endpoint and patch a
flags value.
6. If Gateway Setting RUNTIME_FEATURE_FLAGS == True, the update succeeds
and flag value is resource synced to EDA
7. If Gateway Setting RUNTIME_FEATURE_FLAGS == False, the update fails
with error explaining why.